+++ shopapp-angular/src/app/components/detail-product/detail-product.component.html
@@ -58,8 +58,8 @@
           <div class="product-quantity">
             Quantity:
             <div class="border-wrapper">
-              <button (click)="decreaseQuantity()">-</button>
-              <input type="text" [value]="quantity" (input)="updateQuantity($event)">
+              <button (click)="decreaseQuantity()" id="decreaseIndex">-</button>
+              <input type="text" id="inputNumber" [value]="quantity" (input)="updateQuantity($event)">
               <button (click)="increaseQuantity()" id="increaseIndex">+</button>
             </div>
           </div>
